파티룸 목록 api 추가 및 sql 수정

+++ api/src/main/java/com/pfplaybackend/api/playlist/presentation/dto/PlayListCreateDto.java
@@ -10,7 +10,7 @@ import lombok.Getter;
 @Builder
 public class PlayListCreateDto {
     private final PlayList playList;
-    private final Long orderNumber;
+    private final Integer orderNumber;
     private final String name;
     private final PlayListType type;
     private final User user;
